// packages/datadog-api-client-v2/models/IncidentTeamResponseAttributes.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* The incident team's attributes from a response.
*/
export declare class IncidentTeamResponseAttributes {
    /**
    * Timestamp of when the incident team was created.
    */
    'created'?: Date;
    /**
    * Timestamp of when the incident team was modified.
    */
    'modified'?: Date;
    /**
    * Name of the incident team.
    */
    'name'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): IncidentTeamResponseAttributes;
    static serialize(data: IncidentTeamResponseAttributes): {
        [key: string]: any;
    };
    constructor();
}